Moving indices nav files into search-index folder
