+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_index/indices/indices_nav.scss
